This is the designed version of Starter theme
If you want some styles already applied to your starter theme this is for you. And yes there are some improvements in the layout, the starter is fully from scratch and this version has some styles already applied + layout setup done. In layout header section, posts section and about me are added.